Add Custom ListView for viewing events
Add a custom list adapter and UI for viewing events in the
events list tab.  The UI is not complete, but it provides
a start on how to create these custom listViews.  The next
step will be to encapsulate the custom ListView in a swipe
view so that the user can swipe between past, today, and
future events.