Add support for ViewPagerIndicator
This adds supporting files for Jake Wharton's ViewPagerIndicator
custom UI components.  This is being utilized for custom tab/swipe
fragments for use throughout the app.  I am still experimenting
with these new components, so it is still possible that we decide
not to use them.  This code is licensed under Apache Lincense,
Version 2.0.  Please view more licensing info in any of the
files.