Fix CreateEvent Swipe Null Data Issue
What this change does, is listens for when the viewpager swipe is on
the Invite page. When it is, it will listen for the Create Event button
click. For some reason, we are able to get the data from the Details
page and we should be able to create the query from here. When this is
completed, some of this code can be removed from CreateEventSwipe.java.